Add sources and scripts to generate country.
This is a new approach for managing boundary data in the commons.
The basic components are:
	- Locations of source data in a 'vrt' file.
	- shell script to run ogr2ogr commands to create shapefiles
          and CSVs
	- SQL file with query to create appropriate geometries
          and attribute data

This approach is suited to countries where:
	- Boundary data is in a relativly stable location.
	- The data only requires manipulation of existing attributes
	  and geometries.

Advantages are:
	- As long as the schemas stay the same update can be handled
	  by changing the source.vrt file.
	- All data processing is explicit in the vrt and sql files.
	- No need to store source data

Disadvantages are:
	- If source is taken down we don't have a copy.
	- Some operations are slow - eg. point in polygon look ups.

source.vrt
The VRT file is the gdal 'virtual format' you can read about it here:
https://www.gdal.org/drv_vrt.html. When used as the input file of the
ogr2ogr command all the files referenced within the VRT can be
accessed as layers - this means in the sql they look like tables in a
a db. This reduces processing steps and intermediary files.
In addition reprojection is handled by using 'OGRVRTWarpedLayer' type.

make-country.sh
Is the shell script to make the shp and CSV.

make-country.sql
Is the query used in the ogr2ogr command in make-country.sh. When the
number of features is low attribute data can be hard coded.